+++ cmd/zz_gen_cmd_dnshelp.go
@@ -23,6 +23,7 @@ func allDNSCodes() string {
 		"azure",
 		"bindman",
 		"bluecat",
+		"bunny",
 		"checkdomain",
 		"civo",
 		"clouddns",
@@ -89,11 +90,13 @@ func allDNSCodes() string {
 		"nicmanager",
 		"nifcloud",
 		"njalla",
+		"nodion",
 		"ns1",
 		"oraclecloud",
 		"otc",
 		"ovh",
 		"pdns",
+		"plesk",
 		"porkbun",
 		"rackspace",
 		"regru",
@@ -327,6 +330,25 @@ func displayDNSHelp(w io.Writer, name string) error {
 
 		ew.writeln()
 		ew.writeln(`More information: https://go-acme.github.io/lego/dns/bluecat`)
+
+	case "bunny":
+		// generated from: providers/dns/bunny/bunny.toml
+		ew.writeln(`Configuration for Bunny.`)
+		ew.writeln(`Code:	'bunny'`)
+		ew.writeln(`Since:	'v4.11.0'`)
+		ew.writeln()
+
+		ew.writeln(`Credentials:`)
+		ew.writeln(`	- "BUNNY_API_KEY":	API key`)
+		ew.writeln()
+
+		ew.writeln(`Additional Configuration:`)
+		ew.writeln(`	- "BUNNY_POLLING_INTERVAL":	Time between DNS propagation check`)
+		ew.writeln(`	- "BUNNY_PROPAGATION_TIMEOUT":	Maximum waiting time for DNS propagation`)
+		ew.writeln(`	- "BUNNY_TTL":	The TTL of the TXT record used for the DNS challenge`)
+
+		ew.writeln()
+		ew.writeln(`More information: https://go-acme.github.io/lego/dns/bunny`)
 
 	case "checkdomain":
 		// generated from: providers/dns/checkdomain/checkdomain.toml
@@ -1717,6 +1739,26 @@ func displayDNSHelp(w io.Writer, name string) error {
 		ew.writeln()
 		ew.writeln(`More information: https://go-acme.github.io/lego/dns/njalla`)
 
+	case "nodion":
+		// generated from: providers/dns/nodion/nodion.toml
+		ew.writeln(`Configuration for Nodion.`)
+		ew.writeln(`Code:	'nodion'`)
+		ew.writeln(`Since:	'v4.11.0'`)
+		ew.writeln()
+
+		ew.writeln(`Credentials:`)
+		ew.writeln(`	- "NODION_API_TOKEN":	The API token`)
+		ew.writeln()
+
+		ew.writeln(`Additional Configuration:`)
+		ew.writeln(`	- "NODION_HTTP_TIMEOUT":	API request timeout`)
+		ew.writeln(`	- "NODION_POLLING_INTERVAL":	Time between DNS propagation check`)
+		ew.writeln(`	- "NODION_PROPAGATION_TIMEOUT":	Maximum waiting time for DNS propagation`)
+		ew.writeln(`	- "NODION_TTL":	The TTL of the TXT record used for the DNS challenge`)
+
+		ew.writeln()
+		ew.writeln(`More information: https://go-acme.github.io/lego/dns/nodion`)
+
 	case "ns1":
 		// generated from: providers/dns/ns1/ns1.toml
 		ew.writeln(`Configuration for NS1.`)
@@ -1830,6 +1872,28 @@ func displayDNSHelp(w io.Writer, name string) error {
 
 		ew.writeln()
 		ew.writeln(`More information: https://go-acme.github.io/lego/dns/pdns`)
+
+	case "plesk":
+		// generated from: providers/dns/plesk/plesk.toml
+		ew.writeln(`Configuration for plesk.com.`)
+		ew.writeln(`Code:	'plesk'`)
+		ew.writeln(`Since:	'v4.11.0'`)
+		ew.writeln()
+
+		ew.writeln(`Credentials:`)
+		ew.writeln(`	- "PLESK_PASSWORD":	API password`)
+		ew.writeln(`	- "PLESK_SERVER_BASE_URL":	Base URL of the server (ex: https://plesk.myserver.com:8443)`)
+		ew.writeln(`	- "PLESK_USERNAME":	API username`)
+		ew.writeln()
+
+		ew.writeln(`Additional Configuration:`)
+		ew.writeln(`	- "PLESK_HTTP_TIMEOUT":	API request timeout`)
+		ew.writeln(`	- "PLESK_POLLING_INTERVAL":	Time between DNS propagation check`)
+		ew.writeln(`	- "PLESK_PROPAGATION_TIMEOUT":	Maximum waiting time for DNS propagation`)
+		ew.writeln(`	- "PLESK_TTL":	The TTL of the TXT record used for the DNS challenge`)
+
+		ew.writeln()
+		ew.writeln(`More information: https://go-acme.github.io/lego/dns/plesk`)
 
 	case "porkbun":
 		// generated from: providers/dns/porkbun/porkbun.toml